+++ gamewin.h
@@ -284,7 +284,7 @@ public:
 	void set_std_delay(int msecs)
 		{ std_delay = msecs; }
 	inline Actor *get_npc(long npc_num) const
-		{ return (npc_num >= 0 && npc_num < npcs.size()) ? 
+		{ return (npc_num >= 0 && npc_num < (int)npcs.size()) ? 
 				npcs[npc_num] : 0; }
 	void set_body(int npc_num, Dead_body *body)
 		{ bodies.put(npc_num, body); }
